% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_cox.R
\name{admm_cox}
\alias{admm_cox}
\title{ADMM algorithm for a 3-component mixture Cox regression model}
\usage{
admm_cox(
  t,
  status,
  x,
  index,
  beta_0,
  alpha1_0,
  alpha2_0,
  alpha3_0,
  vlambda1,
  vlambda2,
  eps = 1e-06,
  max = 20
)
}
\arguments{
\item{t}{observed survival time.}

\item{status}{observed censoring indicator.}

\item{x}{covariates.}

\item{index}{component index.}

\item{beta_0}{a vector of the average effect beta.}

\item{alpha1_0}{a vector of the deviance of effects in the 1st component from the average effect.}

\item{alpha2_0}{a vector of the deviance of effects in the 2nd component from the average effect.}

\item{alpha3_0}{a vector of the deviance of effects in the 3rd component from the average effect.}

\item{vlambda1}{a vector of tuning parameters for beta.}

\item{vlambda2}{a vector of tuning parameters for alpha.}

\item{eps}{the convergence criteria.}

\item{max}{maximum number of iterations.}
}
\value{
coefficients estimates of beta and alpha
}
\description{
Update the parameter estimates using ADMM algorithm in EM algorithm
}
