% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admm_linear2.R
\name{admm_linear2}
\alias{admm_linear2}
\title{ADMM algorithm for a 2-component mixture linear regression model}
\usage{
admm_linear2(
  y,
  x,
  w,
  beta_0,
  alpha1_0,
  alpha2_0,
  vlambda1,
  vlambda2,
  eps = 1e-06,
  max = 50
)
}
\arguments{
\item{y}{observed response variable.}

\item{x}{covariates.}

\item{w}{weights.}

\item{beta_0}{a vector of the average effect beta.}

\item{alpha1_0}{a vector of the deviance of effects in the 1st component from the average effect.}

\item{alpha2_0}{a vector of the deviance of effects in the 2nd component from the average effect.}

\item{vlambda1}{a vector of tuning parameters for beta.}

\item{vlambda2}{a vector of tuning parameters for alpha.}

\item{eps}{the convergence criteria.}

\item{max}{maximum number of iterations.}

\item{index}{component index.}
}
\value{
coefficients estimates of beta and alpha
}
\description{
Update the parameter estimates using ADMM algorithm in EM algorithm
}
\keyword{internal}
