% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_admm_linear2.R
\name{gd_phi2}
\alias{gd_phi2}
\title{Internal function for admm_linear2}
\usage{
gd_phi2(phi_0, beta, alpha1, alpha2, beta_t, alpha1_t, alpha2_t)
}
\description{
an internal function to update phi in admm through gradient descent
}
\keyword{internal}
