% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_admm_linear2.R
\name{gd_theta2}
\alias{gd_theta2}
\title{Internal function for admm_linear2}
\usage{
gd_theta2(y, x, w, beta, alpha1, alpha2, phi, beta_t, alpha1_t, alpha2_t)
}
\description{
an internal function to update theta in admm through gradient descent
}
\keyword{internal}
