% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_admm_linear3.R
\name{gd_theta3}
\alias{gd_theta3}
\title{Internal function for admm_linear3}
\usage{
gd_theta3(
  y,
  x,
  w,
  beta,
  alpha1,
  alpha2,
  alpha3,
  phi,
  beta_t,
  alpha1_t,
  alpha2_t,
  alpha3_t
)
}
\description{
an internal function to update theta in admm through gradient descent
}
\keyword{internal}
