% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_admm_cox.R
\name{update_phi_cox}
\alias{update_phi_cox}
\title{Internal function for admm_cox}
\usage{
update_phi_cox(
  phi_0,
  beta,
  alpha1,
  alpha2,
  alpha3,
  beta_t,
  alpha1_t,
  alpha2_t,
  alpha3_t
)
}
\description{
an internal function to update thetat in admm_cox
}
\keyword{internal}
