% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_admm_cox.R
\name{update_theta_cox}
\alias{update_theta_cox}
\title{Internal function for admm_cox}
\usage{
update_theta_cox(
  t,
  status,
  x,
  index,
  beta,
  alpha1,
  alpha2,
  alpha3,
  phi,
  beta_t,
  alpha1_t,
  alpha2_t,
  alpha3_t
)
}
\description{
an internal function to update theta in admm_cox
}
\keyword{internal}
