% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_admm_cox.R
\name{update_thetat_cox}
\alias{update_thetat_cox}
\title{Internal function for admm_cox}
\usage{
update_thetat_cox(beta, alpha1, alpha2, alpha3, phi, vlambda1, vlambda2)
}
\description{
an internal function to update thetat in admm_cox
}
\keyword{internal}
